pyproject.toml
[project]
name = "robot_learning_baselines"
version = "0.0.0"
description = "A set of baseline models with supporting training/evaluation scripts for a number of robot learning datasets (work is still being ported to this repo)."
authors = [
	{ name = "Peter David Fagan", email = "peterdavidfagan@gmail.com" }
]
readme = "README.md"

[tool.poetry]
name = "robot_learning_baselines"
version = "0.0.0"
description = "A set of baseline models with supporting training/evaluation scripts for a number of robot learning datasets (work is still being ported to this repo)."
license = "Apache-2.0"
authors = [
	"Peter David Fagan <peterdavidfagan@gmail.com>"
]
readme = "README.md"
classifiers = [
    "Operating System :: POSIX :: Linux"
]

[tool.setuptools]
py-modules=["rearrangement_benchmark"]

[tool.poetry.dependencies]
python = "3.10.6"
numpy = "^1.16.0"
pillow = "10.0.0"
matplotlib = "^3.7.2"
hydra-core = "^1.3.2"
notebook = "^7.0.6"
ipywidgets = "^8.1.1"
jupyterlab = "^4.0.9"
wandb = "^0.16.1"
argparse = "^1.4.0"
pandas = "^2.2.0"
openpyxl = "^3.1.2"


# Install jax and jaxlib with CUDA support last to prevent cuda version mismatches
jax = "^0.4.23"
jaxlib = {extras=["cuda12_pip"], url = "https://storage.googleapis.com/jax-releases/cuda12/jaxlib-0.4.23+cuda12.cudnn89-cp310-cp310-manylinux2014_x86_64.whl"}
flax = "^0.7.5"
optax = "^0.1.7"
einops = "^0.7.0"
tqdm = "^4.66.1"
onnx = "^1.15.0"
jaxonnxruntime = "^0.3.0"

# linux-only deps
dm-reverb = {version="0.13.0", markers = "sys_platform == 'linux'"}
tensorflow-cpu = {version="^2.14.0", markers = "sys_platform == 'linux'"}
envlogger = {version="^1.1", extras=["tfds"], markers = "sys_platform == 'linux'"}
tensorflow-datasets = "4.9.3"
rlds = {version="^0.1.7", markers = "sys_platform == 'linux'"}

# submodules
multi_modal_transformers = {path="./robot_learning_baselines/model_architectures/multi_modal_transformers", develop=true}
transporter_networks = {path="./robot_learning_baselines/model_architectures/transporter_networks", develop=true}
octo = {path="./robot_learning_baselines/submodules/octo", develop=true}
dlimp = {path="./robot_learning_baselines/submodules/dlimp", develop=true}
distrax = "^0.1.5"

[tool.black]
line-length = 120

[tool.flake8]
max-line-length = 120
